requirements.txt
beautifulsoup4==4.13.4
playwright==1.54.0
polars==1.32.3
protobuf==6.32.0
python_dateutil==2.9.0.post0
Requests==2.32.5
google-cloud-bigquery==3.25.0
configparser==7.2.0
